Refactor merge_zones from compdat to common
